+++ src/components/study-group-detail/calendar/index.ts
@@ -1,3 +1,4 @@
 export { default as ScheduleCalendar } from '@components/study-group-detail/calendar/ScheduleCalendar'
 export { default as CalendarHeader } from '@components/study-group-detail/calendar/CalendarHeader'
 export { default as CalendarDays } from '@components/study-group-detail/calendar/CalendarDays'
+export { default as CalendarBody } from '@components/study-group-detail/calendar/CalendarBody'
